+++ src/promptflow-devkit/promptflow/_sdk/_orchestrator/experiment_orchestrator.py
@@ -53,7 +53,7 @@ from promptflow._sdk._orm.experiment import Experiment as ORMExperiment
 from promptflow._sdk._orm.experiment_node_run import ExperimentNodeRun as ORMExperimentNodeRun
 from promptflow._sdk._orm.orchestrator import Orchestrator as ORMOrchestrator
 from promptflow._sdk._orm.run_info import RunInfo as ORMRunInfo
-from promptflow._sdk._utils import overwrite_null_std_logger
+from promptflow._sdk._utilities.general_utils import overwrite_null_std_logger
 from promptflow._sdk.entities import Run
 from promptflow._sdk.entities._experiment import Experiment, ExperimentTemplate
 from promptflow._sdk.operations import RunOperations